refactor: split unicode map from icon font
