+++ packages/IconFont/index.js
@@ -2,7 +2,8 @@ import React from 'react'
 
 import { toPascalCase } from '../../src/utils/toPascalCase'
 
-import { Icon, unicodeMap } from './styles'
+import { Icon } from './styles'
+import { unicodeMap } from './unicode'
 
 export const IconFont = Icon
