+++ packages/IconFont/styles.js
@@ -2,176 +2,6 @@ import styled, { css } from '@xstyled/styled-components'
 import { th } from '@xstyled/system'
 import { system } from '@welcome-ui/system'
 
-export const unicodeMap = {
-  actions: '\f101',
-  activity: '\f102',
-  add: '\f103',
-  address: '\f104',
-  alert: '\f105',
-  anniversary: '\f106',
-  archive: '\f107',
-  ats: '\f108',
-  attachment: '\f109',
-  avatar: '\f10a',
-  basketball: '\f10b',
-  behance: '\f10c',
-  birthday: '\f10d',
-  bold: '\f10e',
-  book: '\f10f',
-  bulb: '\f110',
-  bulk: '\f111',
-  calcul: '\f112',
-  camera: '\f113',
-  candidate_status: '\f114',
-  celebrate: '\f115',
-  check: '\f116',
-  down: '\f117',
-  left: '\f118',
-  right: '\f119',
-  up: '\f11a',
-  chevron: '\f11b',
-  code: '\f11c',
-  coffee: '\f11d',
-  collapse: '\f11e',
-  comment_outline: '\f11f',
-  comment: '\f120',
-  compass: '\f121',
-  computer: '\f122',
-  contract: '\f123',
-  cover_letter: '\f124',
-  create: '\f125',
-  cross_target: '\f126',
-  cross: '\f127',
-  crown: '\f128',
-  csv: '\f129',
-  date: '\f12a',
-  department: '\f12b',
-  details: '\f12c',
-  dish: '\f12d',
-  drag_and_drop: '\f12e',
-  dribbble: '\f12f',
-  earth: '\f130',
-  edit: '\f131',
-  education_level: '\f132',
-  email_outline: '\f133',
-  email: '\f134',
-  employer_branding: '\f135',
-  expand: '\f136',
-  external_link: '\f137',
-  eye: '\f138',
-  facebook: '\f139',
-  female: '\f13a',
-  doc: '\f13b',
-  docx: '\f13c',
-  eot: '\f13d',
-  gif: '\f13e',
-  jpeg: '\f13f',
-  jpg: '\f140',
-  pdf: '\f141',
-  png: '\f142',
-  svg: '\f143',
-  ttf: '\f144',
-  txt: '\f145',
-  woff: '\f146',
-  woff2: '\f147',
-  xls: '\f148',
-  xlsx: '\f149',
-  zip: '\f14a',
-  filters: '\f14b',
-  flag_pointed: '\f14c',
-  flag: '\f14d',
-  gear: '\f14e',
-  get: '\f14f',
-  github: '\f150',
-  handshake: '\f151',
-  hard_hat: '\f152',
-  hashtag: '\f153',
-  heading_1: '\f154',
-  heading_2: '\f155',
-  heart_outline: '\f156',
-  heart_shiny: '\f157',
-  heart: '\f158',
-  info: '\f159',
-  information: '\f15a',
-  instagram: '\f15b',
-  italic: '\f15c',
-  link: '\f15d',
-  linkedin: '\f15e',
-  location: '\f15f',
-  lock_rounded: '\f160',
-  lock: '\f161',
-  log_out: '\f162',
-  male: '\f163',
-  map: '\f164',
-  marketplace: '\f165',
-  megaphone: '\f166',
-  menu: '\f167',
-  mic: '\f168',
-  microship: '\f169',
-  money: '\f16a',
-  more: '\f16b',
-  mountain: '\f16c',
-  music: '\f16d',
-  negative_outline: '\f16e',
-  negative_straight: '\f16f',
-  negative: '\f170',
-  notification: '\f171',
-  offer_status: '\f172',
-  ordered_list: '\f173',
-  origine: '\f174',
-  paperplane: '\f175',
-  pencil: '\f176',
-  phone_outline: '\f177',
-  phone: '\f178',
-  piechart: '\f179',
-  pin: '\f17a',
-  pizza: '\f17b',
-  play_outline: '\f17c',
-  play: '\f17d',
-  podcast: '\f17e',
-  positive_outline: '\f17f',
-  positive_straight: '\f180',
-  positive: '\f181',
-  print: '\f182',
-  promote: '\f183',
-  question: '\f184',
-  questions: '\f185',
-  quote: '\f186',
-  refuse_candidate: '\f187',
-  remote: '\f188',
-  reset: '\f189',
-  resume: '\f18a',
-  review_outline: '\f18b',
-  review: '\f18c',
-  search: '\f18d',
-  settings: '\f18e',
-  shield: '\f18f',
-  source: '\f190',
-  spanner: '\f191',
-  stackoverflow: '\f192',
-  stackserver: '\f193',
-  star: '\f194',
-  strikethrough: '\f195',
-  suitcase: '\f196',
-  sun: '\f197',
-  tag: '\f198',
-  target: '\f199',
-  thumbup: '\f19a',
-  trash: '\f19b',
-  tree: '\f19c',
-  twitch: '\f19d',
-  twitter: '\f19e',
-  unordered_list: '\f19f',
-  update: '\f1a0',
-  upload: '\f1a1',
-  welcome_home: '\f1a2',
-  welcome_kit: '\f1a3',
-  write: '\f1a4',
-  wttj_icons: '\f1a5',
-  xing: '\f1a6',
-  youtube: '\f1a7'
-}
-
 // stylelint-disable font-family-no-missing-generic-family-keyword
 export const Icon = styled.i(
   ({ name, size = 'md' }) => css`
